+++ packages/taiko-client-rs/crates/whitelist-preconfirmation-driver/Cargo.toml
@@ -6,6 +6,7 @@ publish = false
 license.workspace = true
 
 [dependencies]
+alethia-reth-evm = { workspace = true }
 alethia-reth-primitives = { workspace = true }
 alloy-consensus = { workspace = true }
 alloy-eips = { workspace = true }
